requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
isort==5.13.2
black==24.4.2
